+++ src/main/webapp/include/header.jsp
@@ -89,7 +89,7 @@ if (arr != null) {
 						<li class="li-icon chat">
 							<!-- 채팅 아이콘 -->
 							<div class="icon-area">
-								<a href="${pageContext.request.contextPath }/user/message_list.jsp">
+								<a href="${pageContext.request.contextPath }/member/list.message">
 									<img src="${pageContext.request.contextPath }/resources/img/message.png" alt="채팅">
 									<span class="new-mark">2</span>
 								</a>
